update gem for 1.0 hanami version
