+++ lib/hanami-scaffold.rb
@@ -1,0 +1,1 @@
+require 'hanami/scaffold'
